src/core/project_parser.py
# File: src/core/project_parser.py
"""
Handles parsing of FL Studio project files using PyFLP library.
Extracts clip metadata, positions, and arrangement information.

Filepath: src/core/project_parser.py
"""
from typing import List, Dict
import pyflp

class FLProjectParser:
    def __init__(self, project_path: str):
        self.project_path = project_path
        # TODO: Initialize PyFLP parser
        
    def parse_project(self) -> Dict:
        """Parse FL Studio project and extract metadata"""
        # TODO: Implement project parsing logic
        pass
    
    def extract_arrangements(self) -> List[Dict]:
        """Extract arrangement information"""
        # TODO: Implement arrangement extraction
        pass