tests/test_project_parser.py
# File: tests/test_project_parser.py
"""
Tests for FL Studio project parsing functionality.

Filepath: tests/test_project_parser.py
"""
import pytest
from src.core.project_parser import FLProjectParser

def test_project_loading(sample_project_path):
    """Test loading FL Studio project file"""
    parser = FLProjectParser(sample_project_path)
    # TODO: Add assertions for project loading

def test_arrangement_extraction(sample_project_path):
    """Test extracting arrangements from project"""
    parser = FLProjectParser(sample_project_path)
    # TODO: Add assertions for arrangement extraction

def test_clip_metadata_extraction(sample_project_path):
    """Test extracting clip metadata"""
    parser = FLProjectParser(sample_project_path)
    # TODO: Add assertions for clip metadata

def test_invalid_project_handling():
    """Test handling of invalid project files"""
    # TODO: Add assertions for error handling